mission05/agent/requirements.txt
openai
openai-agents
foundry-local-sdk

